+++ backend/Storage.go
@@ -7,6 +7,7 @@ import (
 	"io/ioutil"
 	"log"
 	"math/rand"
+	"net/http"
 	"os"
 	"time"
 
@@ -108,4 +109,36 @@ func CreateCustomToken(ID string) (string, error) {
 	}
 
 	return token, nil
+}
+
+// CheckAuthorized checks if a user is authorized to make a request
+func CheckAuthorized(w http.ResponseWriter, r *http.Request) bool {
+	token := r.Header.Get("Token")
+	id := r.Header.Get("Identity")
+
+	if token == "" || id == "" {
+		respondWithError(w, Unauthorized, "You are not authorized to make this request")
+		return false
+	}
+
+	user, err := fbClient.Ref("/users/" + id + "/accessToken")
+	if err != nil {
+		fmt.Println("Error fetching user " + id + " from Firebase for authentication")
+		respondWithError(w, FailedDBCall, err.Error())
+		return false
+	}
+
+	var aToken map[string]interface{}
+	if err := user.Value(&aToken); err != nil {
+		fmt.Println("Error fetching value of user " + id + " from Firebase for authentication")
+		respondWithError(w, FailedDBCall, err.Error())
+		return false
+	}
+
+	if aToken["access_token"] == token {
+		return true
+	}
+
+	respondWithError(w, Unauthorized, "You are not authorized to make this request")
+	return false
 }
